% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{barrier}
\alias{barrier}
\title{Example of layer representing dispersal barriers for a species}
\format{
A SpatRaster with 108 rows, 84 columns, 9072 cells:
\describe{
\item{barrier}{barriers are represented with values of 1}
}
}
\description{
A raster layer representing dispersal barriers for a species in a region
where simulations can be performed in examples.
}
\examples{
barrier <- terra::rast(system.file("extdata/barrier.tif",
                                      package = "grinnell"))

terra::plot(barrier)
}
