% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_simulation.R
\name{dispersal_simulation}
\alias{dispersal_simulation}
\title{Helper function to simulate species dispersal processes}
\usage{
dispersal_simulation(data, suit_layers, dispersal_kernel = "normal",
  kernel_spread = 1, max_dispersers = 4, replicates = 50,
  dispersal_events = 50, access_threshold = 5,
  write_replicates = FALSE, output_directory = "M_simulation_results")
}
\arguments{
\item{data}{(character) name of the csv file with all the species occurrences,
columns must be:
species, longitude, latitude.}

\item{suit_layers}{(character) vector of suitability layers to be used for
each simulation step.}

\item{dispersal_kernel}{(character) dispersal kernel (dispersal function) used to
simulate the movement
of the species. Options are: "normal", "log_normal". Default = "normal".}

\item{kernel_spread}{(numeric) Spread of the function ***. Default = 2.}

\item{max_dispersers}{(numeric) maximum number of dispersors that depart from
each colonized pixel. Depending
on suitability this number will decrease in less suitable areas.}

\item{replicates}{(numeric) number of times that the simulation will be repeated.}

\item{dispersal_events}{(numeric) ...}

\item{access_threshold}{(numeric) percentage, . Default = 5.}

\item{write_replicates}{(logical) whether or not to write simulation replicates.
Default = TRUE.}

\item{output_directory}{(character) name of the output directory to be created
in which subdirectories
containing all results will be written. Directory name must include the complete path.}
}
\description{
dispersal_simulation generates a python script that performs the
simulation of the species accessible area (M).
}
