% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_simulationR.R
\name{dispersal_simulationR}
\alias{dispersal_simulationR}
\title{Helper function to simulate species dispersal processes}
\usage{
dispersal_simulationR(data, suit_layers, starting_porportion = 0.5,
                      dispersal_kernel = "normal",
                      kernel_spread = 1, max_dispersers = 4,
                      replicates = 10, dispersal_events = 25,
                      threshold = 5, set_seed = 1,
                      return = "all", write_to_directory = FALSE,
                      write_all_scenarios = FALSE,
                      raster_format = "GTiff", output_directory)
}
\arguments{
\item{data}{data.frame containing geographic coordinates of occurrences of
the species of interest. Columns must be: species, longitude, latitude, in
that order.}

\item{suit_layers}{(character) vector of names of suitability layers to be
used as distinct scenarios. If more than one, the layer names must be ordered
starting from the oldest scenario. Layer names should include parent
directory if needed.}

\item{starting_porportion}{(numeric) proportion of \code{data} to be used as
starting points for the simulation. Default = 0.5. All data is used if a
value of 1 is defined.}

\item{dispersal_kernel}{(character) dispersal kernel (dispersal function)
used to simulate the movement of the species. Options are: "normal",
"log_normal". Default = "normal".}

\item{kernel_spread}{(numeric) standard deviation for the
\code{dispersal_kernel}. Default = 1.}

\item{max_dispersers}{(numeric) maximum number of dispersers that depart from
each colonized pixel. Depending on suitability this number will automatically
decrease in areas with low suitability values. Default = 4.}

\item{replicates}{(numeric) number of times to repeat the simulation
per scenario. Default = 10.}

\item{dispersal_events}{(numeric) number of dispersal events to happen per
scenario. Default = 25.}

\item{threshold}{(numeric) percentage to be considered when excluding
accessed or colonized cells with lower values. Default = 5.}

\item{set_seed}{(numeric) a seed to be used when sampling \code{data}
according to \code{starting_porportion}. Default = 1.}

\item{return}{(character) the results to be returned or written. Options are:
"all", "accessed", "colonized". Default = "all"}

\item{write_to_directory}{(logical) whether to write results in
\code{output_directory}. Default = FALSE.}

\item{write_all_scenarios}{(logical) whether or not to write results
for all scenarios. The default, FALSE, writes only the final results if
\code{write_to_directory} = TRUE.}

\item{raster_format}{(character) format to use for raster layers to be
written. Options are: "GTiff", "EHdr", and "ascii", Default = "GTiff.}

\item{output_directory}{(character) name of the output directory where
results should be written. If this directory does not exist, it will be
created.}
}
\value{
If \code{return} = "all', all elements described below will be returned as a
list, if "accessed" or "colonized" are chosen instead, only the elements
corresponding to either "accessed" or "colonized" areas will be returned.

The list returned contains:
\itemize{
\item a list with a summary of scenarios and parameters used
\item a binary RasterLayer showing accessed = 1 and non-accessed areas = 0
\item a RasterLayer representing mean values of accessibility frequency among
replicates
\item a RasterLayer representing variance among values of accessibility frequency
of all replicates
\item a binary RasterLayer showing colonized = 1 and non-colonized areas = 0
\item a RasterLayer representing mean values of frequency of colonization among
replicates
\item a RasterLayer representing variance among values of frequency of
colonization of all replicates
}

If \code{write_to_directory} is set to TRUE, the elements described above
and raster layers corresponding to all scenarios
(if \code{write_all_scenarios} = TRUE) will be written in
\code{output_directory}.
}
\description{
dispersal_simulationR performs a multistep simulation of species
dispersal to reconstruct areas that have been accessed and/or colonized
based on environmental suitability and user-defined dispersal parameters.
}
\examples{
# data
data("records", package = "grinnell")
suitability <- system.file("extdata/suitability.tif", package = "grinnell")

# simulation
d_s <- dispersal_simulationR(data = records, suit_layers = suitability,
                             replicates = 3, dispersal_events = 5,
                             return = "colonized")
}
