% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_suitability.R
\name{ellipsoid_suitability}
\alias{ellipsoid_suitability}
\title{Suitability model based on Mahalanobis distance to niche centroid}
\usage{
ellipsoid_suitability(data, variables, suitability_threshold = 5,
                      project = FALSE, projection_variables,
                      tolerance = 1e-60)
}
\arguments{
\item{data}{a numerical matrix containing geographic coordinates of species
occurrences to be used; columns must be: longitude and latitude.}

\item{variables}{a SpatRaster of variables representing the environmental
conditions in the area of interest.}

\item{suitability_threshold}{(numeric) value (percentage) to be used as
threshold; default = 5.}

\item{project}{(logical) whether or not to project the model to other
scenario(s). If TRUE, argument \code{projection_variables} must be defined.
Default = FALSE.}

\item{projection_variables}{a SpatRaster (if only one scenario) or named
list of SpatRaster objects (if more than one scenario) with variables
representing the environmental conditions of scenarios to transfer the model
to. Variable names must match between initial and projection scenarios.}

\item{tolerance}{the tolerance for detecting linear dependencies.
Default = 1e-60.}
}
\value{
A list containing:
\itemize{
\item occurrences (geographic coordinates and environmental values)
\item niche centroid
\item covariance matrix
\item proportion of suitable areas in environmental and geographic space
\item suitability SpatRaster if \code{projection_variables} is a SpatRaster or
list of suitability SpatRaster(s) if \code{projection_variables} is a list
}
}
\description{
ellipsoid_suitability produces a model of environmental
suitability based on the environmental Mahalanobis distance to the centriod
of the environmental values characterized by species occurrences.
}
\details{
Distance used for creating the model is Mahalanobis distance. All
values outside the ellipsoid produced using the centroid and covariance
matrix derived from environmental values in \code{data}, and the value in
\code{suitability_threshold} will be zero.

Values in maps (from 0 to 1) can be interpreted as suitability.
}
