% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_simulation.R
\name{forward_simulation}
\alias{forward_simulation}
\title{Simulation of species dispersal to future}
\usage{
forward_simulation(suit_layer, data = NULL, suit_forward = NULL,
                   barriers = NULL, starting_proportion = 0.5,
                   proportion_to_disperse = 1, sampling_rule = "random",
                   dispersal_kernel = "normal",
                   kernel_spread = 1, max_dispersers = 4,
                   dispersal_events = 25, replicates = 10,
                   threshold = 5, set_seed = 1,
                   out_format = "GTiff", output_directory)
}
\arguments{
\item{suit_layer}{(character) name of (local) raster layer containing values
of suitability for the species of interest in the study region over which the
simulation will run. The name of the layer name should include parent
directory if needed.}

\item{data}{(optional) data.frame containing geographic coordinates of
occurrences of the species of interest. Columns must be: "species", "longitude",
"latitude", in that order.}

\item{suit_forward}{(optional) name of (local) raster layer(s) containing
values of suitability for the species of interest in the study region over
which the simulation will run. If more than one, layer names must be
ordered from first to last scenario. These layers must have the same
extent, resolution, number of cells, and projection than \code{suit_layer}.
Layer names should include parent directory if needed.}

\item{barriers}{(optional) RasterLayer representing dispersal barriers for
the species. This layer must have the same extent and projection than
\code{suit_layers}. The only values allowed in this layer are 1 and NA;
1 = barrier. Default = NULL.}

\item{starting_proportion}{(numeric) proportion of \code{data} to be used as
starting points for the simulation. Default = 0.5. All data is used if a
value of 1 is defined.}

\item{proportion_to_disperse}{(numeric) proportion of colonized cells from
which dispersers will start a new dispersal event; default = 1.}

\item{sampling_rule}{(character) rule to be used to sample a
\code{starting_proportion} of \code{data} and a \code{proportion_to_disperse}
from colonized cells to run dispersal simulation steps. Options are: "random"
and "suitability". Using the option "suitability" prioritizes records in
with higher suitability values. Default = "random".}

\item{dispersal_kernel}{(character) dispersal kernel (dispersal function)
used to simulate the movement of the species. Options are: "normal",
"log_normal". Default = "normal".}

\item{kernel_spread}{(numeric) standard deviation for the
\code{dispersal_kernel}. Default = 1.}

\item{max_dispersers}{(numeric) maximum number of dispersers that depart from
each colonized pixel. Depending on suitability this number will automatically
decrease in areas with low suitability values. Default = 4.}

\item{dispersal_events}{(numeric) number of dispersal events to happen per
scenario. A vector of multiple values could be used to define different
dispersal events for distinct scenarios. See details; default = 25.}

\item{replicates}{(numeric) number of times to repeat the simulation
per dispersal event, depending on \code{results_by}. Default = 10.}

\item{threshold}{(numeric) percentage to be considered when excluding
accessed or colonized cells with lower values. Default = 5.}

\item{set_seed}{(numeric) a seed to be used when sampling \code{data}
according to \code{starting_proportion}. Default = 1.}

\item{out_format}{(character) format of raster layers to be written in
\code{output_directory}. Options are "ascii", "GTiff", and "EHdr" = bil.
Default = "GTiff".}

\item{output_directory}{(character) name of the output directory where
results should be written. If this directory does not exist, it will be
created.}
}
\value{
A list containing:
\itemize{
\item occurrences found in suitable areas in the scenario where the simulation
started.
\item all scenarios considered for the simulation
\item a list with the parameters used during the simulation
\item a RasterLayer with values representing the number of the
dispersal event when areas where accessed
\item a RasterLayer with values representing the number of the
dispersal event when areas where colonized
\item if defined, the RasterLayer used in \code{barriers}, else NULL
}

The complete set of results derived from data preparation and the simulation
is written in \code{output_directory}. These results include the ones
mentioned above (except barriers), plus:
\itemize{
\item if needed, a folder containing results from correcting suitability layer(s)
with \code{barriers}
\item other raster layers representing statistics of accessibility:
mean and variance
\item a figure representing accessed and colonized areas per dispersal events,
and the occurrences used for simulation
\item a simple report from the simulation process
}

The number of dispersal events in results is continuous among scenarios. If
10 dispersal events are defined and multiple scenarios exist in
\code{suit_layers}, the first dispersal event in the second scenario will be
number 11.
}
\description{
Simulation of species dispersal to identify areas
that could be accessed and colonized based on environmental
suitability and user-defined dispersal parameters.
}
\details{
Defining a vector of multiple values in \code{dispersal_events} could be
useful when distinct scenarios represent different periods of time, or if
a reduced number of events need to be simulated in the last scenario.
If a vector of values is defined in \code{dispersal_events}, the length of
this vector must match the length of \code{suit_forward} + 1, otherwise,
the first element in \code{dispersal_events} will be used and a warning
message will be printed.
}
\examples{
# data
data("records", package = "grinnell")
suitability <- system.file("extdata/suitability.tif", package = "grinnell")

# simulation current
f_s <- forward_simulation(suit_layer = suitability, data = records,
                          dispersal_kernel = "normal",
                          kernel_spread = 2, max_dispersers = 2,
                          dispersal_events = 15, replicates = 3,
                          output_directory = file.path(tempdir(), "eg_fsim"))

# simulation current and future
\donttest{
suitf <- system.file("extdata/suitability_fut.tif", package = "grinnell")

f_s1 <- forward_simulation(suit_layer = suitability, data = records,
                           suit_forward = suitf, dispersal_kernel = "normal",
                           kernel_spread = 2, max_dispersers = 2,
                           dispersal_events = 15, replicates = 3,
                           output_directory = file.path(tempdir(), "eg_fsim1"))

# simulation current and future using dispersal barriers
barrier <- raster::raster(system.file("extdata/barrier.tif",
                                      package = "grinnell"))

f_s2 <- forward_simulation(suit_layer = suitability, data = records,
                           suit_forward = suitf, barriers = barrier,
                           dispersal_kernel = "normal",
                           kernel_spread = 2, max_dispersers = 2,
                           dispersal_events = 15, replicates = 3,
                           output_directory = file.path(tempdir(), "eg_fsim2"))
}
}
