% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{interpolation}
\alias{interpolation}
\title{Helper function to interpolate and prepare layer names}
\usage{
interpolation(ellipsoid_model, types_clim, pos_scenarios, current, lgm,
  current_suitability, lgm_suitability, slash_type, pca_directory,
  suitability_directory)
}
\description{
Helper function to interpolate and prepare layer names
}
