% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolation_values}
\alias{interpolation_values}
\title{Helper function to generate interpolation values}
\usage{
interpolation_values(simulation_period = 140, transition_to_lgm = 13,
                     stable_lgm = 25, lgm_to_current = 7,
                     stable_current = 25, scenario_span = 3)
}
\arguments{
\item{simulation_period}{(numeric) time in thousands of years for the
complete period of simulation.}

\item{transition_to_lgm}{(numeric) time in thousands of years for the
transition period from current-like (interglacial) to glacial (LGM) climatic
conditions.}

\item{stable_lgm}{(numeric) time in thousands of years for the period when
glacial (LGM) conditions are assumed to be relatively stable.}

\item{lgm_to_current}{(numeric) time in thousands of years for the
transition period from glacial (LGM) to current-like (interglacial) climatic
conditions.}

\item{stable_current}{(numeric) time in thousands of years for the period when
current-like (interglacial) conditions are assumed to be relatively stable.}

\item{scenario_span}{(numeric) time in thousands of years that have to pass
for changing the scenario. Default = 3 (three thousand years).}
}
\value{
A list containing:
\itemize{
\item a vector with proportions of interglacial conditions corresponding to
each interpolation
\item a vector represenitng the position of each interpolation in the previous
vector.
}
}
\description{
Helper function to generate interpolation values
}
