% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{M_preparation}
\alias{M_preparation}
\title{Helper function to prepare M files}
\usage{
M_preparation(directory, pattern = "A_S\\\\d.*c$",
  crs = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{directory}{(character)}

\item{pattern}{(character)}

\item{crs}{(character)}
}
\description{
Helper function to prepare M files
}
