% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_simulation.R
\name{M_simulation}
\alias{M_simulation}
\title{Simulation of species accessible areas (M in BAM)}
\usage{
M_simulation(data, current_variables, project = FALSE,
  projection_variables, scale = TRUE, dispersal_kernel = "normal",
  kernel_spread = 1, max_dispersers = 4, suitability_threshold = 5,
  replicates = 10, dispersal_events = 20, access_threshold = 5,
  simulation_period = 50, stable_lgm = 7, transition_to_lgm = 100,
  lgm_to_current = 7, stable_current = 13, scenario_span = 1,
  barriers = NULL, barrier_format = NULL, write_replicates = FALSE,
  output_directory = "M_simulation", plot = TRUE,
  mask_variables = FALSE, directory_masked = "Variables_M")
}
\arguments{
\item{data}{(character) name of the csv file with all the occurrences used to
run the simulation; columns must be: species, longitude, latitude.}

\item{current_variables}{(character) name of the folder where environmental
variables representing current conditions are. Variables must be in ascii
format and at least two. If \code{project} = TRUE, variables in
\code{projection_variables} must have the same extent than variables in this
folder.}

\item{project}{(logical) whether or not to perform the simulation using
suitabilities that change among scenarios that are interpolated based in
current and past conditions. If TRUE, arguments \code{projection_variables},
\code{simulation_period}, \code{stable_current}, \code{stable_lgm},
\code{transition_to_lgm}, \code{lgm_to_current}, and \code{scenario_span},
need to be defined. Default = FALSE.}

\item{projection_variables}{(character) name of the folder where environmental
variables for representing the Last Glacial Maximum scenario are. Variables
must be in ascii format and they must correspond with variables in
\code{current_variables} (i.e., their name and extent must be the same but
conditions should represent the LGM).}

\item{dispersal_kernel}{(character) dispersal kernel (dispersal function)
used to simulate the movement of the species. Options are: "normal",
"log_normal". Default = "normal".}

\item{kernel_spread}{(numeric) standar deviation for the
\code{dispersal_kernel}. Default = 2.}

\item{max_dispersers}{(numeric) maximum number of dispersors that depart from
each colonized pixel. Depending on suitability this number will decrease in
less suitable areas.}

\item{suitability_threshold}{(numeric) percentage of omission error to define
what is suitable and whai is not. Default = 5.}

\item{replicates}{(numeric) number of times that the complete simulation
process will be repeated.}

\item{dispersal_events}{(numeric) number of dispersal events to take place
during the entire period of simulation if \code{project} = FALSE, or each
\code{scenario_span} if \code{project} = TRUE.}

\item{access_threshold}{(numeric) percentage of frequency values to be
considered as unprobably visited during the process of dispersal. Default = 5.}

\item{simulation_period}{(numeric) time in thousands of years for the complete
simulation.}

\item{stable_lgm}{(numeric) time in thousands of years in which climate is
assumed to be relatively stable in the last maximum glacial scenario.}

\item{transition_to_lgm}{(numeric)}

\item{lgm_to_current}{(numeric)}

\item{stable_current}{(numeric) time in thousands of years in which climate
is assumed to be relatively stable in the current scenario.}

\item{scenario_span}{(numeric) time in thousands of years that have to pass
for changing the scenario in which the simulation will be performed.
Default = 1 (one thousend years).}

\item{barriers}{(character) "optional" name of the folder where the barriers
for the species dispersal are. Barriers must be the same projection than
variables (WGS84 and not planar projection is recommended). This barriers can
be in two formats: ascii format with the same resolution than variables,
or shapefile.}

\item{barrier_format}{(character) if \code{barriers} are defined, format of
the layers to be used. Options are: "ascii" and "shp".}

\item{write_replicates}{(logical) whether or not write M ascii files resulted
of each replicate. Default = FALSE}

\item{output_directory}{(character) name of the output directory to be created
in which subdirectories containing all results will be written. Default =
"M_simulation".}

\item{plot}{(logical) whether or not to plot the species' niche ellipsoid and
the resultant M on the environmental suitability based on the species' niche,
the species occurrences are included as well. Default = TRUE.}

\item{mask_variables}{(logical) whether or not to mask the variables to the
simulated accessible area (M). Default = FALSE}

\item{directory_masked}{(character) name of the folder to be created to save
the masked variables. Default = "M_variables".}
}
\value{
Folder \code{output_directory} conatining the results of the simulation.
These results include: A plot of the M and the occurrences on an environmental
layer map that will return to a graphic device if \code{plot.ellipse} = TRUE;
the M as a shapefile and as a raster layer in ascii format; a folder with ....

If \code{mask_variables} = TRUE, the environmental variables will be masked
to the M and the masked layers will be written in the \code{directory_masked}
folder.
}
\description{
M_simulation generates an area that has been potentially
accessible for a given species during relevant periods of time, derived from
a dispersal simulation process.
}
\details{
When more than three variables are present in the \code{var.folder} directory,
a principal component analysis of them is performed. The the three first
principal components are used then, instead of the complete set of variables,
when creating the suitability layer over which the dispersal simulation is
performed.

If barriers are used, the suitability values in the areas where barriers exist
turn into cero. This is, populations cannot stablish there and dispersal will
be truncated unless the dispersal kernel defined by argumens
\code{dispersal_kernel} and \code{kernel_spread}, allow the species to overpass
the barriers.
}
