% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oscillayer.R
\name{oscillayer}
\alias{oscillayer}
\title{Interpolate from the Last Glacial Maximum (LGM) to earlier time periods as oscillayers}
\usage{
oscillayer(
  current_variables,
  lgm,
  final_period = 120,
  precipitation_var = paste0("bio_", 12:19),
  write_files = FALSE,
  output_dir = NULL,
  overwrite = FALSE,
  progress_bar = TRUE
)
}
\arguments{
\item{current_variables}{(SpatRaster) environmental variables representing
"current" conditions.}

\item{lgm}{(SpatRaster) environmental variables representing the conditions
of the Last Glacial Maximum (LGM). It must have the same variables, variable names,
and resolution as the current_variables.}

\item{final_period}{(numeric) The final period (in thousands of years before
present) for which interpolation is performed, with intervals of 10,000 years.
Valid values range from 30 to 5380 and must be multiples of 10. Default = 120.}

\item{precipitation_var}{(character) names of the variables representing precipitation
(e.g., c("bio_12", "bio_13", etc.)). This is necessary for fixing negative
values in the interpolated precipitation variables.}

\item{write_files}{(logical) whether to write TIF files to disk. Default = FALSE.}

\item{output_dir}{(character) directory to save TIF files if write files. Default = NULL}

\item{overwrite}{(logical) whether to overwrite files in the output directory. Default = FALSE.}

\item{progress_bar}{(logical) whether to show progress bar. Default = TRUE.}
}
\value{
A list of SpatRasters containing the interpolated bioclimatic variables for
each period.
}
\description{
oscillayer generates bioclimatic variables from the Last Glacial Maximum (LGM)
to 5.4 million years ago with an interval of 10,000 years, based on the approach
developed by Gamisch (2019) on Oscillayers. See details and references for more
information.
}
\details{
The interpolated bioclimatic variables are constructed using anomalies between
bioclimatic layers of the present and the Last Glacial Maximum (LGM), scaled
relative to the Plio-Pleistocene global mean temperature curve derived from benthic
stable oxygen isotope ratios (see \code{data("scalings")}, \code{?scalings} and references
for details).The function allows interpolation across 538 scenarios, spanning
from the LGM (~20,000 years before present) to 5.4 million years ago.
}
\references{
Gamisch, A. (2019). Oscillayers: A dataset for the study of climatic
oscillations over Plio‐Pleistocene time‐scales at high spatial‐tempora
l resolution. Global Ecology and Biogeography, 28(11), 1552-1560.
}
