% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_raster.R
\name{pca_raster}
\alias{pca_raster}
\title{Principal componens for raster layers and projections}
\usage{
pca_raster(variables, in_format = "ascii", out_format = "ascii",
  scale = TRUE, project = FALSE, projection_variables,
  return_back = FALSE, n_pcs, output_directory = "PCA_results")
}
\arguments{
\item{variables}{(character) name of the folder where environmental variables
are.}

\item{in_format}{(character) format of variables in \code{variables}. Options
are "ascii", "GTiff", and "EHdr" = bil. Default = "ascii".}

\item{out_format}{(character) format of variables to be written in distinct
sets inside \code{output_directory}. Options are "ascii", "GTiff", and
"EHdr" = bil. Default = "ascii".}

\item{scale}{(logical) wheter or not to scale variables while performing
principal component analyses.}

\item{project}{(logical) whether or not to project the species niche to other
scenario(s). If TRUE, argument \code{projection_variables} needs to be defined.
Default = FALSE.}

\item{projection_variables}{(character or RasterStack) if character, name of
the folder where subfolders with environmental variables of scenarios for
projections are (useful if multiple projections are needed). If RasterStack,
object containing stacked variables of only one projection scenario.
Variables must correspond with variables in \code{variables} (i.e., their name
must correspond but they should represent conditions in other scenario).}

\item{return_back}{(logical) whether or not return raster layers of principal
components to the R environment in a list with other results.}

\item{n_pcs}{(numeric) number of principal components to be returned as rasters.
By default all principal components are returned as RasterLayers.}

\item{output_directory}{(character) name of the folder to be created to save
the results of the analyses. Default = "PCA_results".}
}
\value{
A list containing PCA summary and PCA loadings as matrices;
if \code{return_back} = TRUE, one or multiple (if projected)
RasterStacks of principal components are returned additionally.

All results are written in \code{output_directory}.
}
\description{
pca_raster performs a principal component analysis with a set of
variables and produces raster layers of them. If needed the pricipal
components are projected to other scenarios.
}
\details{
If \code{scale} = TRUE, variables are centered to cero and scaled using
\code{\link[base]{scale}}.
}
