% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{records}
\alias{records}
\title{Occurrence records for an example species}
\format{
A data frame with 46 rows and 3 columns.
\describe{
\item{species}{character, species scientific name.}
\item{longitude}{numeric, longitude values.}
\item{latitude}{numeric, latitude values.}
}
}
\source{
\url{https://www.gbif.org/}
}
\usage{
records
}
\description{
A dataset containing geographic coordinates of an example
species.
}
\examples{
data("records", package = "grinnell")
head(records)
}
\keyword{datasets}
