% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{save_Mplot}
\alias{save_Mplot}
\title{Helper function to save and plot M figure}
\usage{
save_Mplot(ellipsoid_model, variables, suitability_layer, M_polygon,
  size_proportion = 0.55, output_directory, plot)
}
\description{
Helper function to save and plot M figure
}
