% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{save_nicheplot}
\alias{save_nicheplot}
\title{Helper function to save and plot niche elllipsoid figure}
\usage{
save_nicheplot(ellipsoid_model, suitability_threshold, variables,
  size_proportion = 0.55, output_directory, plot)
}
\description{
Helper function to save and plot niche elllipsoid figure
}
