% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{suitability}
\alias{suitability}
\title{Example of layer representing environmental suitability for a species}
\format{
A SpatRaster with 108 rows, 84 columns, 9072 cells:
\describe{
\item{suitability}{values from low = 0 to high = 1}
}
}
\description{
A raster layer representing distinct levels of suitability for a species
in a region where simulations can be performed in examples.
}
\examples{
suitability <- terra::rast(system.file("extdata/suitability.tif",
                                          package = "grinnell"))

terra::plot(suitability)
}
