% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{variables_lgm}
\alias{variables_lgm}
\title{Example of variables to represent current climate condition in a region}
\format{
A SpatRaster with 108 rows, 84 columns, 9072 cells, and 6 layers:
\describe{
\item{Temperature}{temperature, in Celsius degrees times 10.}
\item{Precipitation}{precipitation, in milimeters.}
}
}
\source{
\url{https://www.worldclim.org/data/index.html}
}
\description{
A dataset containing raster layers of climatic variables representing Last
Glacial Maximum conditions for a region where simulations can be performed
in examples.
}
\examples{
variables_lgm <- terra::rast(system.file("extdata/variables_lgm.tif",
                                           package = "grinnell"))

terra::plot(variables_lgm[[1]])
}
