% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal_simulation.R
\name{dispersal_simulation}
\alias{dispersal_simulation}
\title{Helper function to simulate species dispersal processes}
\usage{
dispersal_simulation(data, suit_layers, dispersal_kernel = "normal",
                     kernel_spread = 1, max_dispersers = 4,
                     replicates = 10, dispersal_events = 25,
                     access_threshold = 5, output_directory)
}
\arguments{
\item{data}{(character) name of the csv file with all the species occurrences,
columns must be: species, longitude, latitude.}

\item{suit_layers}{(character) vector of names of suitability layers to be
used as distinct scenarios. If more than one, the layer names must be ordered
from past to current. Layer names should include parent directory if needed.}

\item{dispersal_kernel}{(character) dispersal kernel (dispersal function)
used to simulate the movement of the species. Options are: "normal",
"log_normal". Default = "normal".}

\item{kernel_spread}{(numeric) standard deviation for the
\code{dispersal_kernel}. Default = 1.}

\item{max_dispersers}{(numeric) maximum number of dispersers that depart from
each colonized pixel. Depending on suitability this number will automatically
decrease in areas with low suitability values. Default = 4.}

\item{replicates}{(numeric) number of times to repeat the simulation
per scenario. Default = 10.}

\item{dispersal_events}{(numeric) number of dispersal events to happen per
scenario. Default = 25.}

\item{access_threshold}{(numeric) percentage to be considered when excluding
accessed cells with lower values. Default = 5.}

\item{output_directory}{(character) name of the output directory where
results should be written. If this directory does not exist, it will be
created.}
}
\description{
dispersal_simulation generates a python script to perform a
multistep simulation of species dispersal to reconstruct areas that have been
accessed based on environmental suitability and user-defined dispersal parameters.
}
