% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_suitability.R
\name{ellipsoid_suitability}
\alias{ellipsoid_suitability}
\title{Suitability model based on niche centroid distance}
\usage{
ellipsoid_suitability(data, variables, suitability_threshold = 5,
  project = FALSE, projection_variables, tolerance = 1e-60)
}
\arguments{
\item{data}{a numerical matrix containing geographic coordinates of the species
occurrences to be used; columns must be: longitude and latitude.}

\item{variables}{a RasterStack of variables representing the environmental
conditions in the area on which the models will be created.}

\item{suitability_threshold}{(numeric) value from 0 to 100 that will be used as
threshold (E); default = 5.}

\item{project}{(logical) whether or not to project the species niche to other
scenario(s). If TRUE, argument \code{projection_variables} needs to be defined.
Default = FALSE.}

\item{projection_variables}{a RasterStack (if only one scenario) or named list of
RasterStacks (if more than one scenario) of variables representing the
environmental conditions of the scenario at which the species niche will be
transferred. Variables names must correspond between initial and projection
scenarios.}
}
\value{
A list containing: the occurrences (geographic coordinates and environmental
values), the corrdinates of the niche centroid, the covariance matrix, the
proportion of non-suitable areas as measured in the environmental (unique
environmental conditions) and the geographic space (all conditions), and a
suitability RasterLayer (if \code{projection_variables} is a RasterStack) or
a suitability RasterStack (if \code{projection_variables} is a list of
RasterStacks).
}
\description{
ellipsoid_suitability produces a model of environmental
suitability based on the Mahalanobis distance of each environmental value
in a given set of variables, to the centriod of the environmental values
characterized by the occurrences of a given species.
}
\details{
Distance used for creating the model is Mahalanobis distance. All
values outside the ellipsoid produced by the distance measured from the
centroid limit of the occurrences exluding a percentage given by the
\code{suitability_threshold}, will have cero suitability values.

Values in maps (from 0 to 1) are interpreted as suitability. For better
visualization use a color palette based on percetage of data with distinct
values.
}
