% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolation}
\alias{interpolation}
\title{Helper function to interpolate and prepare layer names}
\usage{
interpolation(ellipsoid_model, suitability_threshold = 5,
              interpolation_values, barriers = NULL,
              current_variables, projection_variables,
              current_suitability, lgm_suitability,
              out_format = "GTiff", output_directory)
}
\arguments{
\item{ellipsoid_model}{object produced with the function
\code{ellipsoid_suitability}.}

\item{suitability_threshold}{(numeric) percentage of omission error to define
what is not suitable. Default = 5.}

\item{interpolation_values}{list of results from using the function
\code{\link{interpolation_values}}.}

\item{barriers}{SpatRaster representing dispersal barriers for the species.
This layer must have the same extent and projection than
\code{current_variables}. The only values allowed in this layer are 0 and 1;
0 = barrier, 1 = all other areas. Default = NULL.}

\item{current_variables}{SpatRaster of environmental variables representing
"current" conditions (interglacial).}

\item{projection_variables}{SpatRaster of environmental variables
representing the "Last Glacial Maximum" scenario. Variable names, projection
and extent of these layers must be the same than those in
\code{current_variables}.}

\item{current_suitability}{(character) name of the file representing values
of suitability for the interglacial (current) scenario.}

\item{lgm_suitability}{(character) name of the file representing values of
suitability for the glacial (LGM) scenario.}

\item{out_format}{(character) format of layers to be written in
\code{output_directory}. Options are "ascii" and "GTiff".
Default = "GTiff".}

\item{output_directory}{(character) name of the folder where results will be
written. This directory must exist.}
}
\value{
A vector with names of files corresponding to each interpolation scenario.

All layers resulting from interpolation will be written in
\code{output_directory}.
}
\description{
Helper function to interpolate and prepare layer names
}
