% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{interpolation_values}
\alias{interpolation_values}
\title{Helper function to generate interpolation values}
\usage{
interpolation_values(transition_to_lgm, stable_lgm, lgm_to_current,
  stable_current, simulation_period, scenario_span)
}
\description{
Helper function to generate interpolation values
}
