% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_raster.R
\name{pca_raster}
\alias{pca_raster}
\title{Principal component analysis for raster layers}
\usage{
pca_raster(variables, in_format = NULL, scale = TRUE, center = TRUE,
           n_pcs = NULL, project = FALSE, projection_variables,
           return_projection = FALSE, write_to_directory = FALSE,
           out_format = "GTiff", output_directory, overwrite = FALSE)
}
\arguments{
\item{variables}{SpatRaster or character. If SpatRaster, stack of layers to
be used; if character, name of the folder where environmental variables
are.}

\item{in_format}{(character) if class of \code{variables} = "character",
format of variables in such a folder. Options are "ascii", "GTiff", and
"EHdr" = bil. Default = NULL.}

\item{scale}{(logical) whether or not to scale variables while performing
principal component analyses.}

\item{center}{(logical) whether or not to center variables while performing
principal component analyses.}

\item{n_pcs}{(numeric) number of principal components to be returned as
raster layers. By default all principal components are returned.}

\item{project}{(logical) whether or not to project the species niche to other
scenario(s). If TRUE, argument \code{projection_variables} must be defined.
Default = FALSE.}

\item{projection_variables}{character or SpatRaster. If character, name of
the folder where one or more sub-directories containing variables of distinct
environmental scenarios are (useful if multiple projections are needed).
If SpatRaster, stack of variables representing one scenario for projection.
Variable names must correspond with those of \code{variables} (i.e., their
names must match).}

\item{return_projection}{(logical) whether to return raster PCs for
projection scenario(s) as part of the resulting list. Default = FALSE.}

\item{write_to_directory}{(logical) whether to write results in
\code{output_directory}. Default = FALSE.}

\item{out_format}{(character) format of layers to be written in
\code{output_directory}. Options are "ascii", "GTiff", and "EHdr" = bil.
Default = "GTiff".}

\item{output_directory}{(character) name of the folder to be created to save
results.}

\item{overwrite}{(logical) whether or not to overwrite the
\code{output_directory} if it already exists. Default = FALSE.}
}
\value{
A list containing:
\itemize{
\item PCA results
\item SpatRaster of principal components for \code{variables}
\item if return_projection = TRUE, list of SpatRaster objects with principal
components for other scenarios. NULL if \code{project} = FALSE.
}
}
\description{
pca_raster performs a principal component analysis with a set of
variables and returns results including PCs as raster layers. If needed,
principal components can be projected to other scenarios.
}
