% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_suitability.R
\name{predict_esuitability}
\alias{predict_esuitability}
\title{Predict suitability based on niche centroid distances}
\usage{
predict_esuitability(ellipsoid_model, centroid, covariance_matrix,
  variables, suitability_threshold = 5, tolerance = 1e-60)
}
\arguments{
\item{ellipsoid_model}{object produced with the function
\code{ellipsoid_suitability}. If defined, arguments \code{centroid} and
\code{covariance_matrix} are not necessary.}

\item{centroid}{centroid to wich distance will be measured in \code{variables}.
Length of centroid must correspond with number of layers in \code{variables}.
Ignored if \code{ellipsoid_model} is defined.}

\item{covariance_matrix}{covariance matrix to be used in measuring the
Mahalanobis distances. Ignored if \code{ellipsoid_model} is defined.}

\item{variables}{a RasterStack of variables representing the environmental
conditions in the area on which the models will be created.}

\item{suitability_threshold}{(numeric) value from 0 to 100 that will be used
as threshold (E); default = 5.}

\item{tolerance}{the tolerance for detecting linear dependencies.
Default = 1e-60.}
}
\description{
predict_esuitability predicts a RasterLayer of environmental
suitability based on Mahalanobis distances to a niche centroid and a
covariance matrix.
}
