% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{suitability_fut}
\alias{suitability_fut}
\title{Example of layer representing future environmental suitability for a species}
\format{
A RasterLayer with 108 rows, 84 columns, 9072 cells:
\describe{
\item{suitability}{values from low = 0 to high = 1}
}
}
\description{
A raster layer representing distinct levels of future suitability for a
species in a region where simulations can be performed in examples.
}
\examples{
suitability_fut <- raster::raster(system.file("extdata/suitability_fut.tif",
                                              package = "grinnell"))

raster::plot(suitability_fut)
}
