% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\name{variables}
\alias{variables}
\title{Example of variables to represent current climate condition in a region}
\format{
A SpatRaster with 108 rows, 84 columns, 9072 cells, and 6 layers:
\describe{
\item{Temperature}{temperature, in Celsius degrees times 10.}
\item{Precipitation}{precipitation, in milimeters.}
}
}
\source{
\url{https://www.worldclim.org/data/index.html}
}
\description{
A dataset containing raster layers of climatic variables for a region where
simulations can be performed in examples.
}
\examples{
variables <- terra::rast(system.file("extdata/variables.tif",
                                       package = "grinnell"))

terra::plot(variables[[1]])
}
